pyproject.toml
[project]
name = "github-rag-assistant"
version = "0.1.0"
description = "A RAG system for querying GitHub repositories"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "PyGithub>=2.1.0",
    "python-dotenv>=1.0.0",
    "GitPython>=3.1.0",
    "tiktoken>=0.5.0",
    "tomli>=2.0.0",
    "pinecone>=5.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"